pyproject.toml
[tool.poetry]
name = "onyx"
version = "0.12.2"
description = "API for pathogen metadata."
authors = ["Thomas Brier <t.brier@outlook.com>"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
django-filter = "^23.5"
django-simple-history = "^3.5.0"
pydantic = "^2.6.2"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
coverage = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
